pyproject.toml
[tool.poetry]
name = "gen-ai-photoshoots"
version = "0.1.0"
description = ""
authors = ["Theskrtnerd <tvbbd2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.5.4"
python-dotenv = "^1.0.1"
datasets = "^2.19.1"
torch = "^2.3.0"
diffusers = "^0.27.2"
transformers = "^4.41.0"
bitsandbytes = "^0.43.1"
accelerate = "^0.30.1"
ftfy = "^6.2.0"
torchvision = "^0.18.0"
pyproject-flake8 = "^7.0.0"

[tool.flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
